ARM: MitySOM: testfixture: Fix ADC initialization.
With the pinmuxing initialization from 3.1, the ain signals could not be
found by the system. Changed to initialization similar to that of the
devkit's touchscreen.

The new ADC driver caused the major number of ttyO0, which is used for
the console, to change. The old major number was hard-coded in the
initramfs, so devtmpfs is now needed to find ttyO0's new major number.