ARM: MitySOM: testfixture: Switch ECC method to software BCH8.
The SOM's nand and the test fixture's nand cannot both use the ELM. Let
the SOM use the ELM, which means the test fixture cannot use hardware
BCH.

The test fixture's nand's datasheet specifies the minimum ECC
required as 4-bit, so Hamming codes should not be used. Use software
BCH8 as it is the largest amount of ECC that can fit in the test
fixture's nand's OOB area.